experiments/bootloader/requirements.txt
pyserial==3.4
RPi.GPIO==0.6.3
